+++ include/vosp/contracts.hpp
@@ -4,6 +4,7 @@
 
 #include <vosp/contracts/error.hpp>
 #include <vosp/contracts/logging.hpp>
+#include <vosp/contracts/telemetry.hpp>
 #include <vosp/contracts/version.hpp>
 
 #ifndef VOSP_NAMESPACE_FACADE_DEFINED